+++ entomb/exceptions.py
@@ -1,3 +1,7 @@
+class FileIntegrityError(Exception):
+    """Raise when there is a mismatch between a file and its data files."""
+
+
 class GetAttributeError(Exception):
     """Raise when the immutable attribute cannot be accessed."""